Remove pointless integration build tag from shadertools
